etl-pipeline/pyproject.toml
[tool.poetry]
name = "etl-pipeline"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^1.4.2"
psycopg2 = "^2.9.3"
prefect = "2.*"
python-dotenv = "^0.19.0"


[tool.poetry.dev-dependencies]
pytest = "^6.2.4"


[tool.poetry.scripts]
main = "etl_pipeline.main:main"